+++ Backend/cryptoCurrency/model/src/main/java/ar/utn/tacs/model/coin/Coin.java
@@ -1,13 +1,14 @@
 package ar.utn.tacs.model.coin;
 
 import java.math.BigDecimal;
+import java.math.BigInteger;
 
 import org.codehaus.jackson.annotate.JsonIgnoreProperties;
 
 @JsonIgnoreProperties(value = {"id"})
 public class Coin {
 
-	private Long id;
+	private BigInteger id;
 	private String name;
 	private String ticker;
 	private BigDecimal valueInDollars = new BigDecimal(0);
@@ -15,7 +16,7 @@ public class Coin {
 	public Coin() {
 	}
 
-	public Coin(Long id, String name, String ticker) {
+	public Coin(BigInteger id, String name, String ticker) {
 		this.id = id;
 		this.name = name;
 		this.ticker = ticker;
@@ -31,11 +32,11 @@ public class Coin {
 		return this.getTicker().equals(other.getTicker());
 	}
 	
-	public Long getId() {
+	public BigInteger getId() {
 		return id;
 	}
 
-	public void setId(Long id) {
+	public void setId(BigInteger id) {
 		this.id = id;
 	}
